+++ packages/npm-package-json-lint-config/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 4.25.0 (2023-08-16)
+
+### Enhancement
+
+-   Updated `npm-package-json-lint` peer dependency to require v6.0.0 [#53636](https://github.com/WordPress/gutenberg/pull/53636)
+
 ## 4.24.0 (2023-08-10)
 
 ## 4.23.0 (2023-07-20)